+++ README.md
@@ -1,6 +1,2 @@
 # 2024_1_los_pinguinos
 Tinder frontend repository created by los pinguinos me la van a mascar team
-
-Команда los pinguinos me la van a masca
-Сервис для поиска партнера по совместным спортивным занятиям «SportBro»
-1. Что за сервис? Для чего он нужен? Какие проблемы и задачи пользователей он решает?
